agents/billing/requirements.txt
transformers==4.30.2
torch==2.0.1
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
redis==4.6.0
pika==1.3.2
python-dotenv==1.0.0
requests==2.31.0